requirements.txt
browser_use==0.2.5
Flask==3.1.1
gradio==5.33.0
langchain_openai==0.3.19
pydantic==2.11.5
python-dotenv==1.1.0
Requests==2.32.3
swarms==7.8.4
